% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drip.R
\name{drip}
\alias{drip}
\title{Add a dataset to a puddle.}
\usage{
drip(puddle_db, FETCH_FUN, data_name, description)
}
\arguments{
\item{puddle_db}{A connection to a puddle sqlite database.}

\item{FETCH_FUN}{A zero-argument function that fetches raw data, such as a csv file, xml, or json, and returns it.}

\item{data_name}{A name for the dataset in the puddle. Must be unique in the puddle.}

\item{description}{Descriptive text about the dataset.}
}
\value{
The puddle_db connection, ready for thenext thing you want to do.
}
\description{
Add a dataset to a puddle.
}
