% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaze.R
\name{gaze}
\alias{gaze}
\title{Gaze into the puddle datasets}
\usage{
gaze(puddle_db)
}
\arguments{
\item{puddle_db}{A connection to a puddle database.}
}
\description{
Review all the datasets in a particular puddle.
}
