% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connection.R
\name{get_puddle_connection}
\alias{get_puddle_connection}
\title{Get a connection the puddle's SQLite database.}
\usage{
get_puddle_connection(puddle)
}
\arguments{
\item{puddle}{SQLite connection string to connect to (or create) a puddle database.}
}
\description{
Get a connection the puddle's SQLite database.
}
