% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{PUDDLE_TABLE_NAME}
\alias{PUDDLE_TABLE_NAME}
\title{Internal Utilities}
\format{
An object of class \code{character} of length 1.
}
\usage{
PUDDLE_TABLE_NAME
}
\description{
Internal Utilities
}
\keyword{datasets}
